requirements.txt
pillow==12.1.0
PyMuPDF==1.26.7
